MINOR: Extract HeartbeatRequestState from heartbeat request managers
The AbstractHeartbeatRequestManager and the
StreamsGroupHeartbeatRequestManager, both use the
HeartbeatRequestState to track the state of the heartbeat requests.
Both heartbeat request managers have an implementation of
HeartbeatRequestState as inner class.
To deduplicate code this commit extracts the HeartbeatRequestState
so that the same code can be used by both heartbeat request manager.