MINOR: Downgrade log messages of swallowed exception to warnings
Exceptions that are caught during closing a task dirty are
swallowed. The corresponding log messages are on error level
which is misleading since the exceptions do not cause any
special handling or crash.

This commit downgrades the log messages of swallowed exceptions
to warnings and explains in the log messages that the exceptions
are swallowed.