KAFKA-18913[WIP]: Removal of code guarded by the negation of _state.updater.enabled_ flag
Removed code of methods StreamThread#initializeAndRestorePhase and #initializeAndRestorePhase and related assertions in the tests, since they were only used in the negation flow.

Furthermore, with the following commit the dependency on the _state.updater.enabled_ flag has been removed from the StreamThread#runOnceWithoutProcessingThreads method.

Wanted to verify the changes done are in fact the correct ones from this commit, and the changes to completely remove the _state.updater.enabled_ flag will be done in future commits by taking the reviews into consideration.