KAFKA-18736: Handle errors in the Streams group heartbeat request manager
This commit adds error handling to the Streams heartbeat request manager.

Errors can occur while sending a heartbeat request and when a response
with an error code that is not NONE is received.

Some errors are handled explicitly to recover from them or to log
specific messages. All the others are handled as fatal errors.