Add FunctionalInterface Annotation to most of the interfaces primarily used for lambdas
A lot of the Kafka Stream documentation makes liberal use of Java
lambdas, which is fine. The FunctionalInterface annotation isn't
strictly required for Java to do this, as Java will automatically
coerce, however this breaks the lambda syntax in Clojure 1.12, as it
needs  the FunctionalInterface in order to coerce a Clojure function
into a Java lambda.

As far as I can tell, there's no reason *not* to have the
FunctionalInterface annotation in all the places that regularly use the
lambda syntax, and this annotation would make it more clear.