KAFKA-15931: Cancel RemoteLogReader gracefully
Backports f24945b519005c0bc7a28db2db7aae6cec158927 to 4.0

Instead of reopening the transaction index, it cancels the
RemoteFetchTask without interrupting it--avoiding to close the
TransactionIndex channel.

This will lead to complete the execution of the remote fetch but
ignoring
the results. Given that this is considered a rare case, we could live
with this. If it becomes a performance issue, it could be optimized.