MINOR: Change to allow building on Java 24
Upgrade Scala to 2.13.16, in order to get a Java 24-compatible ASM version.

Upgrade Gradle to 8.14, which is the first Java 24 compatible version.

Upgrade spotbugs to 4.9.4. This is the minimum version that's both
Java 24 compatible, and not affected by https://github.com/spotbugs/spotbugs/issues/3320

Since the new spotbugs version introduces some extra checks, this commit
excludes those warnings, without looking at whether they are real issues
or false positives.

Fix gradle wrapper.

The download-the-wrapper bit was inserted before the CLASSPATH line
in the wrapper, because that referenced the jar. That line no longer
mentions the jar, because the wrapper is referenced via -jar instead of -cp
now.

This means the download was placed in the middle of a java invocation
command, breaking the script.

It's been moved to be before the call to exec. It seems fair to assume
that the script will only exec once, and that the exec call is going to
call the wrapper.