MINOR: Deduplicate topics of a topology for authorization check
With the new Streams rebalance protocol, the Streams client sends
the topology with the used topics to the broker for initialization.
For the initialization the broker needs to describe the topics in
the topology and consequently the Streams application needs to be
authorized to describe the topics.

The broker checks the authorization by filtering the topics in the
topology by authorization. This filtering implicitly deduplicates
the topics of the topology if they appear multiple times in the
topology send to the brokers. After that the broker compares the
size of the authorized topics with the topics in the topology. If
the authorized topics are less than the topics in the topology
a TOPIC_AUTHORIZATION_FAILED error is returned.

In Streams a topology that is sent to the brokers likely has
duplicate topics because a repartition topic appears as a sink
for one subtopology and as a source for another subtopology.

This commit deduplicates the topics of a topology before
the verification of the authorization.