MINOR: Use readable interface to parse response
This is the Response side of Request  refactoring PR in commit: 56d1dc1b6e12a8a325c91749485aa03bce17c7f1

The generated ReequestData classes take Readable as input to parse the Response. However,
the individual Response objects take ByteBuffer as input and thus convert them to Readable
using `new ByteBufferAccessor` call. This call is in all the Respose classes that create
boilderplate code in high tens of these classes.

This PR changes the parse method of all the Respnose class to take Readable instead so that
no such conversion is needed. The clients are the responsible for calling the parse method
with appropriate object which is majority of the cases is already a Readable.

There were couple of places in code where `position` method is used, which is not present
in the Readable interface, so a new `ReadableBuf` interface is created to add these methods.
This allows us to not retreive ByteBuffer from Readable to use the `position` method.

Rest of the changes are all refactoring with no logical change.