KAFKA-18913: Removing changelogReader from ProcessorStateManager, ActiveTaskCreator, StandbyTaskCreator and updates to TaskManager class
Removed the ChangelogReader from the ProcessorStateManager, ActiveTaskCreator, and StandbyTaskCreator and related test cases. This was based on a review in the PR, since the registering with ChangelogReader is done in the old code flow where the state updater is disabled.

Furthermore, the TaskManager class was also updated. And the update includes the removal of code where the condition (stateUpdater == null) is false, since the stateUpdater is not null by default.