Remove shutdown hook in KafkaYammerMetrics (#528)
LI-TICKET: LIKAFKA-64484

- Description

Today, if a Kafka broker is stuck at safety shutdown check during controlled shutdown because of NOT_ENOUGH_REPLICAS, the metrics will be missing for the whole time. This is because Kafka registered a shutdown hook in KafkaYammerMetrics which will call jmxReporter.shutdown() and unregister all its mbean sensors.

To fix the issue, we should explicitly control the metrics shutdown logic. With this change, we shutdown the metrics reporter after everything is shutdown. In the case of controlled shutdown being stuck, we should still see the metrics because the reporter is shutdown at the last step.