KAFKA-19118: Enable KIP-1071 in StandbyTaskCreationIntegrationTest
Enable KIP-1071 parameter in StandbyTaskCreationIntegrationTest.

Required a fix: In ChangelogTopic.setup, we actually need to return
both the source-topic (optmized) and the non-source-topic changelog
topics, since otherwise we will not find the partition number later on.

Extended EmbeddedKafkaCluster to set the number of standby replicas
dynamically for the group. We need to initialize it to one for
the integration test to go through.