Catch exceptions for LogDirFailureHandler thread during handleLogDirF… (#530)
LITICKET=ACTIONITEM-5825

-Description

We recently found an issue that Kafka didn't shutdown itself when there is bad disk issue detected. The current assumption is Kafka detected the issue but the LogDirFailureHandler thread was not working correctly. Unfortunately we don't have enough logs or heapdump to identify what exactly happens, so our best bet is to catch the exceptions for that thread to avoid it happens in the future. We also add a new metric to notify when such exception is thrown, which usually indicate some critical issues.

A more aggressive strategy is to exit the program whenever detecting offline log dir, which might make sense for LinkedIn setup. I don't see huge risk of doing it, but we may still want to start from being a bit conservative. So I decided to add metrics to indicate when a server detects failed disks, so at least our monitoring systems could detect it and potentially taking further actions.