KAFKA-19192: Old bootstrap.metadata files cause problems with post-KAFKA-18601 servers
Old bootstrap.metadata files cause problems with post-KAFKA-18601 servers. When the server tries to
read the bootstrap.metadata file, it will fail if the metadata.version is older than 3.3-IV3
(feature level 7). This causes problems when these clusters are upgraded.

This PR makes it possible to represent older MVs in BootstrapMetadata objects without causing an
exception. An exception is thrown only if we attempt to access the BootstrapMetadata. This ensures
that only the code path in which we start wtih an empty metadata log checks that the metadata
version is 7 or newer.