KAFKA-18905 Flaky-test fixes
As a result of KAFKA-18905 the ReassignPartitionsTest will often have
flaky-test failures which are unrelated to the actual reassignment of
partitions.

The goal of this test is to reassign partitions and make sure that
there is no data-loss. Duplication is not great but it's less
critical.

While we await for KAFKA-18905 to be fixed we can ignore test failures
due to duplicates to reduce flakiness.

Two changes in this test make flakiness much less likely:
- setting retries=5 means that the failure mode encountered in
KAFKA-18905 does not occur. If this issue is fixed then a duplicate
record will be consumed.
- adding new configuration to stop for duplicate consumption.