KAFKA-15370: Support Participation in 2PC (KIP-939) (K/N)
Support keepPreparedTxn=true on transaction coordinator.  The ongoing
transaction needs to be preserved (and its producerId / epoch needs to
be preserved as well so that we can send transaction markers), but the
client-facing producerId / epoch needs to move forward, so now we have 2
id / epoch pairs that need to be maintained during the ongoing and
prepare states.

To simplify state machine implementation, the transaction coordinator
eagerly rejects requests during the prepareCommit / prepareAbort states
with CONCURRENT_TRANSACTIONS error - the client would get this error
anyway and we avoid handing a complex combinaton of conditions that try
to distinguish between retries and invalid requests.