KAFKA-19202: Enable KIP-1071 in streams_eos_test
Enable next system test with KIP-1071.

Some of the validation inside the test does not make sense for KIP-1071.
This is because in KIP-1071, if a member leaves or joins the group,
not all members may enter a REBALANCING state. For streams groups,
we only include a minor sleep. However, this simplification should
not break the test - it can just lead to going too quickly through
the various test states.

Once KAFKA-19271 is implemented, we may change the test app to
print a log line whenever the member epoch is bumped, which is
the only way a member can "indirectly" observe that other members
are rebalancing.