KAFKA-19044: Handle tasks that are not present in the current topology
A heartbeat might be sent to the group coordinator, claiming to own tasks that we do
not know about. We need some logic to handle those requests. In KIP-1071, we propose
to return `INVALID_REQUEST` error whenever this happens, effectively letting the
clients crash.  

This behavior will, however, make topology updates impossible. Bruno Cadonna proposed
to only check that owned tasks match our set of expected tasks if the topology epochs
between the group and the client match. The aim of this change is to implement a
check and a behavior for the first version of the protocol, which is to always
return `INVALID_REQUEST` if an unknown task is sent to the group coordinator.

We can relax this constraint once we allow topology updating with topology epochs. 

To efficiently check this whenever we receive a heartbeat containing
tasks, we precompute the number of tasks for each subtopology. This also
benefits the performance of the assignor.