KAFKA-18913: Cleanup State Updater on Stream Thread creation failure
When an Exception is thrown while a StreamThread is created, a State Updater Thread is left hanging since it is created before the StreamThread constructor is called.

Handled this from this commit after wrapping the code after the creation of the StateUpdater in the StreamThread#create by using a try catch block and manually shutting down the StateUpdater if an exception is thrown before a StreamThread is properly created.