MINOR: Make InternalStreamsBuilder#addGraphNode public
Currently, building the topology with custom graph nodes fails with a NPE because the GraphNode#buildPriority field is null. We can set this value to an empiric one (like taking the one from the previous node), but the best is to reuse the same method as all the other nodes.