KAFKA-19372: StreamsGroup should not be subscribed to a topic when it is empty
We should behave more like a consumer group and make sure to not be subscribed to the input topics anymore when the last member leaves the group. We don't do this right now because our topology is still initialized even after the last member leaves the group.

This will allow:
* Offsets to expire and be cleaned up.
* Offsets to be deleted through admin API calls.