KAFKA-19422: Deflake streams_application_upgrade_test
In this upgrade test, applications sometimes crash before the upgrade,
so it's actually triggering a bug in several older versions (2.x and
possibly others). It seems to be a rare race condition that has been
happening since 2022. Since we are not going to roll out a patch
release for Kafka Streams 2.x, we should just allow applications
to crash before the upgrade.