KAFKA-19429: Deflake streams_smoke_test, again
It looks like we are checking for properties that are not guaranteed
under at_least_once, for example, exact counting (not allowing for
overcounting).

This change relaxes the validation constraint:

The TAGG topic contains effectively count-by-count results. So for
example, if we have the input without duplication

0 -> 1,2,3 we will get in TAGG 3 -> 1, since 1 key had 3 values.

with duplication:

0 -> 1,1,2,3 we will get in TAGG 4 -> 1, since 1 key had 4 values.

This makes the result difficult to compare. Since we run the smoke
test also with Exactly_Once, I propose to disable validation off
TAGG under ALOS.

Similarly, the topic AVG may overcount or undercount. The test case
is extremely similar to DIF, both performing a join and two streams,
the only difference being the mathematical operation performed, so
we can also disable this validation under ALOS with minimal loss of
coverage.

Finally, the change fixes a bug that would throw a NPE when
validation of a windowed stream would fail.