KAFKA-19444: Add back JoinGroup v0 & v1
This fixes librdkafka older than the recently released 2.11.0
with Kerberos authentication and Apache Kafka 4.x.

Even though this is a bug in librdkafka, a key goal of KIP-896 is
not to break the popular client libraries listed in it. Adding
back JoinGroup v0 & v1 is a very small change and worth it from
that perspective.

The cost of adding JoinGroup v0-v1 back is pretty low and hence I suggest we do that.