KAFKA-19842: Fix flaky KafkaStreamsTelemetryIntegrationTest
The new "streams" protocol behaves slightly different to the "classic" protocol, and thus we need to update the test to avoid race conditions.
In particular, the first call to `poll()` won't "block" and return after task assignment completed, but return early without a task assignment, and only a consective rebalance will assign tasks.