MINOR: Run CI with Java 24
This commit updates CI to test against Java 24 instead of Java 23 which is
EOL.

Due to Spotbugs not having released version 4.9.4 yet, we can't run Spotbugs
on Java 24. Instead, we are choosing to run Spotbugs, and the rest of
the compile and validate build step, on Java 17 for now.

Once 4.9.4 has released, we will switch to using Java 24 for this.

Exclude spotbugs from the run-tests gradle action. Spotbugs is already
being run once in the build by "compile and validate", there is no reason
to run it again as part of executing tests.