KAFKA-19596: Improve visibility when topic auto-creation fails
Log a warning for each topic that failed to be created as a result of
an automatic creation.

Previously, at the default log level, you could only see logs that the
broker was attempting to autocreate topics. If the creation failed it
was logged at debug. This makes it more visible.

Signed-off-by: Robert Young <robertyoungnz@gmail.com>