KAFKA-19661: Streams groups sometimes describe as NOT_READY when STABLE
Streams groups sometimes describe as NOT_READY when STABLE. That is,
the group is configured and all topics exist, but when you use
LIST_GROUP and STREAMS_GROUP_DESCRIBE, the group will show up as not
ready.

The root cause seems to be that #19802 moved the creation of the soft
state configured topology from the replay path to the heartbeat. This
way, LIST_GROUP and STREAMS_GROUP_DESCRIBE, which show the snapshot of
the last committed offset, may not show the configured topology, because
the configured topology that is created in the heartbeat is "thrown away",
and the new group is recreated on the replay-path.

We can work around this issue by storing a reference to the streams
group objected created in the heartbeat immediately in the
GroupMetadataManager. This is a snapshottable collection, so if
we fail to write the corresponding records to the consumer offset
topic, this addition of the streams group will be reverted.