KAFKA-19694: Trigger StreamsRebalanceListener in Consumer.close
In the consumer, we invoke the consumer rebalance onPartitionRevoked or onPartitionLost
callbacks, when the consumer closes. The point is that the application may want to
commit, or wipe the state if we are closing unsuccessfully.

In the StreamsRebalanceListener, we did not implement this behavior, which means when
closing the consumer we may lose some progress, and in the worst case also miss that
we have to wipe our local state state since we got fenced.

In this PR we implement StreamsRebalanceListenerInvoker, very similarly
to ConsumerRebalanceListenerInvoker and invoke it in Consumer.close.