KAFKA-19730: StreamsGroupDescribe result is missing topology when topology not configured
In the streams group heartbeat, we validate the topology set for the
group against the topic metadata, to generate the "configured topology"
which has a specific number of partitions for each topic.

In streams group describe, we use the configured topology to expose this
information to the user. However, we leave the topology information
as null in the streams group describe response, if the topology is not
configured. This triggers a null-pointer exception in the admin client
implementation.

Instead, we should expose the unconfigured topology when the configured
topology is not available, which will still expose useful information.