MINOR: Fix StreamsRebalanceListenerInvoker
StreamsRebalanceListenerInvoker was implemented to match the
behavior of ConsumerRebalanceListenerInvoker, however
StreamsRebalanceListener has a subtly different interface than
ConsumerRebalanceListener - it does not throw exceptions, but
returns it as an Optional.

In the interest of consistency, this change fixes this mismatch
by changing the StreamsRebalanceListener interface to behave more
like the ConsumerRebalanceListener - throwing exceptions directly.

In another minor fix, the StreamsRebalanceListenerInvoker is
changed to simply skip callback execution instead of throwing
an IllegalStateException when no streamRebalanceListener is
defined. This can happen when the consumer is closed before
Consumer.subscribe is called.