KAFKA-19779: Add partitions to offset commit validation [1/N]
This adds code to pass committed partitions into
ClassicGroup, StreamsGroup and ConsumerGroup. These are not
used in validation yet, this is just preparation for
later PRs.

The partitions are passed as a supplier, so that we only
instantiate the collection if we are not trying to commit
with the latest member epoch.

We also add variations of streams groups tests for the
tests in OffsetMetadataManagerTest, which previously only
tested the consumer group case. Right now the commit logic
is exactly the same, but we will diverge in the following
PRs.