KAFKA-19779: Introduce source topic to subtopology map [2/N]
For validating offset commits based on partitions, we need be able
to efficiently find the subtopology from a given source topic
(user input topic or repartition topic).

We introduce this precomputed map in `StreamsTopology`, where it
is generated upon construction.

We can use the precomputed map in two places in
StreamsGroup.isSuscribedToTopic we can used it to validate
efficiently if we are still subscribed to a certain topic.

Using `StreamTopology` here instead of `ConfiguredTopology` is
an improvement in itself, because configured topology is soft
state and not entirely reliable to always be there.