MINOR: Clean up coordinator runtime metrics isolation tests
When we fixed KAFKA-19445 to prevent sharing of metrics between the
group coordinator and share coordinator, we added an isolation test for
each metric. This approach is unwieldy and it is very easy to forget to
add isolation tests for new metrics. Replace the isolation tests with
a single test that automatically collects the list of sensors and
metrics from two different CoordinatorRuntimeMetricsImpl instances and
checks them for overlap.