MINOR: Bytes lexicographic comparator could use compiler builtin
The `Arrays.compare` method has a vectorized compiler intrinsic available
and should offer better performance than a hand-rolled loop.

The JDK implementation is at least able to compare 8 bytes
at a time instead of 1 byte at a time, and avoids repeated bounds-checking.

This operation is performance-sensitive if you use in-memory stores, since
it is used to navigate the TreeMap backing the store.