KAFKA-19974: Do not bump member epoch when exiting UNREVOKED_PARTITIONS
Previously, when transitioning from UNREVOKED_PARTITIONS to
UNREVOKED_PARTITIONS, we bumped the member epoch by 1. After the change
for KAFKA-19431, we could also enter UNREVOKED_PARTITIONS due to
unsubscribing from a topic. This introduced a new class of epoch bumps,
which allowed a partition to be owned by different members at the same
epoch (though not at the same time). We would like to preserve the
invariant that partitions can only be owned by a single member at any
given epoch, so we remove the partition bump entirely.

After this patch, when transitioning from UNREVOKED_PARTITIONS to
UNREVOKED_PARTITIONS, the member epoch is not bumped. When transitioning
from UNREVOKED_PARTIIONS to any other state, the member epoch is updated
to match the target assignment epoch.