KAFKA-19862: Group coordinator loading may fail when there is concurrent compaction (#20907)
When the group coordinator loads with concurrent compaction, the record
unassigning a partition/task can be missed which leads to an
IllegalStateException being thrown, however the records self-resolve by
the time loading is finished. This PR makes the change to log a warning
and proceed with loading instead of throwing an IllegalStateException.

Testing: Added unit tests for Consumer and Streams groups to ensure this
change does not impact loading success.

Reviewers: Sean Quah <squah@confluent.io>, Lucas Brutschy
 <lbrutschy@confluent.io>, David Jacot <djacot@confluent.io>