KAFKA-19994: TaskManager may not close all tasks on task timeouts
When a TimeoutException occurs while trying to put multiple active tasks back
into running, we will add the timed out task back to the state updater, so
that we retry it.

However, if we run into a Task timeout (failing to make progress for a long
time), we will rethrow a StreamsException wrapping the TimeoutException we
have drained multiple tasks from the state updater, they will be lost, and
not added back to the state updater, and therefore not be closed correctly.
The task directories remain locked, causing issues trying to replace the
stream thread.