KAFKA-19026: addition of KIP-1252 compatibility flag
With the introduction of KIP-1252, Kafka brokers running in
Zookeeper mode, can invoke an AlterConfigPolicy in a way that
is consistent with KRaft mode.

By default the flag is not enabled, so exixting behavior is preserved.

The unit test checks the behavior in KRaft mode and in Zookeeper mode
with both the flag enabled and disabled.